refactor(search): couple supersession-index test to production predicate
Round-7 review pointed out the EXPLAIN QUERY PLAN regression test ran
against a hand-rolled `SELECT 1 FROM edges e WHERE e.kind = 'updates'
AND e.dst = ?`, not the predicate `do_search_keyword` actually emits.
That left a real false-negative: SQLite's partial-index proof rule
recognises literal-equality conjuncts (`e.kind = 'updates'`) but does
not handle equivalent rewrites such as `e.kind IN ('updates')` or a
parameterised `e.kind = ?`. Either rewrite would silently regress
keyword search to a full edges scan per FTS hit while the toy EQP test
still passed.

Extract the supersession `NOT EXISTS (...)` body as
`SUPERSESSION_NOT_EXISTS_CLAUSE`, a `pub(crate)` const used verbatim by
the production SQL builder. Move the regression test next to the const
as a unit test that EXPLAINs that exact string. The test now also
fails for any rewrite of the production predicate that loses partial-
index applicability — the original goal of the regression guard.

The exact-DDL and `pragma_index_xinfo` shape checks remain.
