+++ crates/cairn-store-sqlite/src/store/search.rs
@@ -50,6 +50,26 @@ use crate::store::{SqliteMemoryStore, current_unix_ms};
 /// see consistent paging semantics across read verbs.
 const SEARCH_LIMIT_MAX: usize = 1000;
 
+/// Records-latest supersession predicate, written so `SQLite`'s partial-index
+/// proof rule can match it against migration 0012's
+/// `INDEX edges_updates_dst_idx ON edges(dst) WHERE kind = 'updates'`.
+///
+/// Why this exact form:
+///
+/// `SQLite` uses a partial index for a query only when it can prove the
+/// query's `WHERE` clause implies the index's `WHERE` clause. The proof
+/// engine handles literal-equality conjuncts (`e.kind = 'updates'`) but
+/// does *not* recognise equivalent rewrites such as `e.kind IN ('updates')`
+/// or a parameterised `e.kind = ?`. Either rewrite would silently
+/// regress the supersession check to a full `edges` scan per FTS hit
+/// (brief §5.1).
+///
+/// The string is shared verbatim with the EXPLAIN-QUERY-PLAN regression
+/// test below so any change here must continue to plan as a `SEARCH`
+/// against the partial index.
+pub(crate) const SUPERSESSION_NOT_EXISTS_CLAUSE: &str = "NOT EXISTS ( SELECT 1 FROM edges e \
+                  WHERE e.kind = 'updates' AND e.dst = r.record_id )";
+
 /// Tokens shown in the FTS5 `snippet()` highlight. Hard-coded so the wire
 /// shape is stable across callers — the verb layer can re-render if it
 /// needs different markers.
@@ -300,11 +320,9 @@ fn build_search_query(
          ) fts ON fts.rowid = r.rowid \
          WHERE r.active = 1 \
            AND r.tombstoned = 0 \
-           AND NOT EXISTS ( \
-                 SELECT 1 FROM edges e \
-                  WHERE e.kind = 'updates' AND e.dst = r.record_id \
-           )",
+           AND ",
     );
+    sql.push_str(SUPERSESSION_NOT_EXISTS_CLAUSE);
     params.push(SqlVal::Text(SNIPPET_OPEN.to_owned()));
     params.push(SqlVal::Text(SNIPPET_CLOSE.to_owned()));
     params.push(SqlVal::Text(SNIPPET_ELLIPSIS.to_owned()));
@@ -433,5 +451,99 @@ trait IntoTokio {
 impl IntoTokio for StoreError {
     fn into_tokio(self) -> tokio_rusqlite::Error {
         tokio_rusqlite::Error::Other(Box::new(self))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    //! Index-coupling tests for the supersession predicate.
+    //!
+    //! These guards live next to the SQL builder so they reference the
+    //! exact constant production uses — rather than a hand-written copy
+    //! that could silently drift. They cover both the index *shape*
+    //! (key column, partial predicate, name) and the *use* of the index
+    //! by the production predicate string under EXPLAIN QUERY PLAN.
+    //!
+    //! The motivating concern (round-7 review): equivalent rewrites such
+    //! as `e.kind IN ('updates')` or a parameterised `e.kind = ?` defeat
+    //! `SQLite`'s partial-index proof rule even though they look correct
+    //! to a human reader. Driving EXPLAIN from
+    //! [`SUPERSESSION_NOT_EXISTS_CLAUSE`] means any such rewrite to the
+    //! production string will surface here as a `SCAN edges` plan.
+    use super::SUPERSESSION_NOT_EXISTS_CLAUSE;
+    use crate::open_in_memory_sync;
+
+    /// Asserts the partial index is shaped correctly *and* used by the
+    /// production supersession clause.
+    ///
+    /// Three independent checks together close the loopholes round-5,
+    /// round-6 and round-7 reviewers identified:
+    /// 1. `pragma_index_xinfo` confirms the leading key column is `dst`.
+    /// 2. `sqlite_schema.sql` matches the canonical partial-index DDL
+    ///    exactly, so any broadened predicate (e.g. adding `OR kind =
+    ///    'mentions'`) fails the test.
+    /// 3. EXPLAIN QUERY PLAN of the *production* predicate string
+    ///    reports `SEARCH ... USING INDEX edges_updates_dst_idx`. This
+    ///    catches both predicate rewrites that lose proof equivalence
+    ///    and accidental removal of the partial index.
+    #[test]
+    fn supersession_clause_uses_partial_index() {
+        let conn = open_in_memory_sync().expect("open");
+
+        // (1) Leading indexed column is `dst`.
+        let mut xinfo = conn
+            .prepare("SELECT name, key FROM pragma_index_xinfo('edges_updates_dst_idx')")
+            .expect("prepare index_xinfo");
+        let key_cols: Vec<(String, i64)> = xinfo
+            .query_map([], |r| Ok((r.get::<_, String>(0)?, r.get::<_, i64>(1)?)))
+            .expect("query index_xinfo")
+            .filter_map(Result::ok)
+            .collect();
+        let leading_key: Vec<&str> = key_cols
+            .iter()
+            .filter(|(_, key)| *key == 1)
+            .map(|(name, _)| name.as_str())
+            .collect();
+        assert_eq!(
+            leading_key,
+            vec!["dst"],
+            "edges_updates_dst_idx must key on dst (got {key_cols:?})",
+        );
+
+        // (2) Partial predicate is exactly the supersession-only one.
+        let ddl: String = conn
+            .query_row(
+                "SELECT sql FROM sqlite_schema \
+                 WHERE type = 'index' AND name = 'edges_updates_dst_idx'",
+                [],
+                |r| r.get(0),
+            )
+            .expect("read index DDL");
+        let normalized: String = ddl.split_whitespace().collect::<Vec<_>>().join(" ");
+        assert_eq!(
+            normalized, "CREATE INDEX edges_updates_dst_idx ON edges(dst) WHERE kind = 'updates'",
+            "broadened or rewritten DDL would re-introduce edges-table bloat. Got: {ddl}",
+        );
+
+        // (3) EXPLAIN the *production* clause. We wrap it in a minimal
+        // outer query that supplies the `r.record_id` referent the
+        // clause closes over, then assert SQLite picks the partial index.
+        let explain_sql = format!(
+            "EXPLAIN QUERY PLAN \
+             SELECT 1 FROM records r WHERE {SUPERSESSION_NOT_EXISTS_CLAUSE}"
+        );
+        let mut stmt = conn.prepare(&explain_sql).expect("prepare EXPLAIN");
+        let plan: Vec<String> = stmt
+            .query_map([], |r| r.get::<_, String>(3))
+            .expect("query plan")
+            .filter_map(Result::ok)
+            .collect();
+        let joined = plan.join("\n");
+        assert!(
+            joined.contains("SEARCH") && joined.contains("edges_updates_dst_idx"),
+            "production supersession clause must SEARCH the partial index. \
+             A regression to e.g. `kind IN ('updates')` would surface here. \
+             Got plan:\n{joined}",
+        );
     }
 }
